pyproject.toml
[project]
name = "local-tool-ai"
version = "0.1.0"
description = "Local tool-calling agent powered by LM Studio"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.30.0",
    "rich>=13.7.0",
    "python-dotenv>=1.0.0",
    "typer>=0.12.0",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "pymupdf4llm>=0.0.17",
    "python-docx>=1.1.0",
    "prompt_toolkit>=3.0.0"
]

[project.scripts]
ai-agent = "local_tool_ai.cli:main"
main = "local_tool_ai.main:run"
server = "local_tool_ai.server:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["local_tool_ai"]

[dependency-groups]
dev = [
    "pytest>=8.2.0",
    "pytest-mock>=3.14.0"
]
lint = [
    "ruff>=0.4.0"
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
